% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eval.R
\name{agg_evals}
\alias{agg_evals}
\title{Aggregation function for eval_splits results}
\usage{
agg_evals(df_evl, var, maximize)
}
\arguments{
\item{var}{Additional grouping variable for results aggregation, e.g. may be variable 'added' or 'removed' during forward or backward selection.}

\item{maximize}{Logical value indicating if optimal result is largest (TRUE) or smallest result (FALSE)}

\item{df_eval}{Results from one or more eval_splits calls.}
}
\value{
tibble with aggregated (averaged) training and validation errors
\describe{
\item{ch_selection}{Column ch_selection from eval_splits call}
\item{var}{Columns passed as var parameter}
\item{selection}{List column of selections, that is from eval_splits call}
\item{m}{Size of aggregated paritions}
\item{mean_train}{Averaged eval_train results, ignoring NAs}
\item{mean_validation}{Averaged eval_validation results, ignoring NAs}
\item{mean_bias}{Averaged bias results, ignoring NAs}
}
}
\description{
Aggregates results from eval_splits call by considering additional grouping variables and
whether a maximization or minimization is required. Flags optimal selection across all results.
}
