% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eval.R
\name{best_selection}
\alias{best_selection}
\title{Helper function to extract best selections}
\usage{
best_selection(agg)
}
\arguments{
\item{agg}{Result from agg_evals call}
}
\value{
selection column for rows flagged as opt (=TRUE)
}
\description{
Helper function to extract best selections
}
