% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_histogram.R
\name{bins_ucv}
\alias{bins_ucv}
\title{Helper function to determine the optimal number of breaks for a histogram}
\usage{
bins_ucv(x)
}
\arguments{
\item{x}{A numeric vector of variables for which the set of breaks should be computed.}
}
\value{
A numeric vector of cut-points points that will define an optimal histogram density estimator for the sample.
}
\description{
This function determines the optimal set of breaks for a histogram for numerical sample
by leave-one-out cross-validation minimizing the empirical risk.
}
