% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_construction.R
\name{box_cox_binomial}
\alias{box_cox_binomial}
\title{Box-Cox transform for binomial models}
\usage{
box_cox_binomial(dat, resp, vars, lambda = seq(-2, +2, 0.1))
}
\arguments{
\item{dat}{A data.frame or tibble comprising the data.}

\item{resp}{A lhs term for the formula. Must be a binary variable (0/1-coded).}

\item{vars}{Character vector of variables for which the Box-Cox transformation should be computed.}

\item{lambda}{Real numeric vector of lambda values to evaluate.}
}
\value{
list with two elements
\describe{
\item{data}{A dataset comprising additional columns with transformed variables, if they improve Normality.}
\item{transformations}{A list of lists with elements for variable, transformed variable, the Box-Cox results, and the transforms.}
}
}
\description{
Determines Box-Cox transformed variables for a binary outcome modelling.
}
