% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_construction.R
\name{box_cox_regression}
\alias{box_cox_regression}
\title{Function to generate Box-Cox transformations for Regression}
\usage{
box_cox_regression(dat, resp, vars, lambda = seq(-2, +2, 0.1))
}
\arguments{
\item{dat}{A data.frame or tibble as dataset.}

\item{resp}{Character string for lhs of the formula.}

\item{vars}{Character vector with variables to transform.}

\item{lambda}{Real vector of lambda coefficients to evaluate for the Box-Cox transform.}
}
\value{
list with two elements
\describe{
\item{data}{A dataset comprising additional columns with transformed variables, if they improve Normality.}
\item{transformations}{A list of lists with elements for variable, transformed variable, the Box-Cox results, and the formula terms.}
}
}
\description{
Applies the Box-Cox transformation for set of variables to improve a regression models. Only
univariate models are considered.
}
