% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_construction.R
\name{check_multimodality}
\alias{check_multimodality}
\title{A function to evaluate continues variables for multi-modiality using Gaussian-Mixture Models}
\usage{
check_multimodality(
  dat,
  resp,
  vars,
  n_comp = 5,
  m_fits = 25,
  min_fits_converged = 24
)
}
\arguments{
\item{dat}{A data.frame or tibble}

\item{resp}{A response term for the lhs of a formula}

\item{vars}{A character vector of variables to secreen}

\item{n_comp}{Integer determining the maximum number of components to check}

\item{m_fits}{Each GMM is fitted multiple times to avoid local minima. m_fits determines how many GMMs are generated each time.}

\item{min_prop_converged}{Minimal proportion of GMM that need to converge before considering their AIC for model selection.}
}
\value{
list with two elements
\describe{
\item{data}{A dataset comprising additional columns with transformed variables, if they improve Normality.}
\item{transformations}{A list of check_aic results for each variable.}
}
}
\description{
The function receives a list of variables and first performs a Gaussian-Mixture Model (GMM)
model selection using the Akaike Information Criterion (AIC) to determine the number of components.
If this process indicates more than one componet, it determines cut-points between the individual 
mixture distributions and generates a transformed categorical variable based on those.
}
\examples{
library( ggplot2 )
library( flexmix )
resp <- "resp"
vars <- grep( "the_|rnd", colnames(toy_data), value=TRUE )

toy_data \%>\%
  ggplot( aes( x=the_multi, y=..density.. ) ) +
  geom_histogram( bins = 100, alpha=0.5 ) +
  geom_density( bw = "ucv" ) 

mumo <- check_multimodality( dat = toy_data, resp = resp, vars = c("the_multi","rnd01","rnd02"),n_comp = 3, m_fits = 25,  min_fits_converged = 20 )

mumo$transforms$the_multi$aic_aggregate
mumo$transforms$the_multi$best_model
parameters(mumo$transforms$the_multi$best_model)
prior(mumo$transforms$the_multi$best_model)
}
