% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_rank.R
\name{game_rank}
\alias{game_rank}
\title{GameRank - A maximum-likelihood based wrapper for variable selection}
\usage{
game_rank(
  dat,
  resp,
  vars,
  fn_train = fn_train_binomial,
  fn_eval = fn_eval_binomial,
  m = NULL,
  team_size = 5L,
  rounds = 50L,
  min_matches_per_var = 10L,
  opt_method = "BFGS",
  max_iter = as.integer(1e+08),
  maximize = TRUE,
  ...
)
}
\arguments{
\item{dat}{Is either a data.frame or tibble with the full dataset used for selection}

\item{resp}{Character string defining response variable (e.g. either a variable name or an formula expression, like 'Surv(time,censored)', used by fn_train and fn_eval)}

\item{vars}{Character strings with variables or variable expressions. Subselections will be fed into fn_train and fn_eval functions.}

\item{team_size}{Selection sizes that are evaluated during match phase against each other}

\item{rounds}{Number of rounds each pair of selections is evaluated on randomly selected train/eval sets}

\item{min_matches_per_var}{Minimum number of matches each variable needs to be part of before the match phase can end}

\item{opt_method}{Should be either 'BFGS' or 'CG', the latter in case Hessian-based optimization is infeasible for the match data}

\item{max_iter}{Maximum number of rounds of the match phase and for group rank model fit.}
}
\value{
List with selection results [TBD]
}
\description{
GameRank - A maximum-likelihood based wrapper for variable selection
}
