% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_rank_helper.R
\name{ll_gr_grad_helper}
\alias{ll_gr_grad_helper}
\title{Define group rank gradient for negative log-likelihood function}
\usage{
ll_gr_grad_helper(vs, res_matches)
}
\description{
Define group rank gradient for negative log-likelihood function
}
\examples{
require(GameRank)
require(numDeriv)
# Some example match data frame
res_matches <- as.data.frame( matrix( c(
1,4, +1, +1, -1, -1,  0,  0,  0,
3,2, +1, +1,  0, -1, -1,  0,  0,
0,5, +1, -1, -1, +1,  0,  0,  0,
5,0, +1, +1,  0,  0,  0, -1, -1,
2,3, -1, +1,  0, +1,  0,  0, -1
),
nrow = 5, ncol = 9, byrow = TRUE,
dimnames = list( NULL, c("n.pos","n.neg", sprintf( "f\%d", 1:7 ) ) ) ) )
res_matches

vs <- c( 0.1, 0.2, 0.3, -0.2, -0.1, 0.2, -0.2 )
vs <- runif( 7 )
names(vs) <- colnames(res_matches)[-c(1,2)]

ll_gr_helper( vs, res_matches )

numericDeriv( quote( ll_gr_helper( vs, res_matches ) ), theta = "vs", rho = as.environment( list( vs=vs, res_matches = res_matches, ll_gr_helper = ll_gr_helper  ) ) )
numDeriv::grad( func = function(vs) ll_gr_helper( vs, res_matches ), x = vs )
ll_gr_grad_helper( vs, res_matches )

}
