% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_rank_helper.R
\name{ll_gr_helper}
\alias{ll_gr_helper}
\title{Evaluating Group Rank Maximum Likelihood estimator}
\usage{
ll_gr_helper(vs, res_matches)
}
\description{
Define group rank negative log-likelihood function from
Huang et al., 2008, p.10, Eq.31
}
