% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reporting.R
\name{render_std_template}
\alias{render_std_template}
\title{Render a standard template report for a given object. (Internal function)}
\usage{
render_std_template(
  template_name,
  output_dir = NULL,
  params,
  output_file = NULL,
  envir = new.env(),
  ...
)
}
\arguments{
\item{template_name}{File name of the standard template}

\item{output_dir}{Output directory to which to render}

\item{params}{The Rmd parameter list. Needs to match the expected parameter definitions of the Rmd template.}

\item{output_file}{Output filename (with or without suffix). Default NULL. If not NULL, it is appended to output_dir}

\item{...}{Further parameters passed to rmarkdown::render(...) function.}
}
\description{
Render a standard template report for a given object. (Internal function)
}
