% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ml_auto_impute.R
\name{rewrite_formula}
\alias{rewrite_formula}
\title{Helper function for Maximum-Likelihood Estimation of Imputation Values}
\usage{
rewrite_formula(fo, dat)
}
\arguments{
\item{fo}{Formula specifying lhs ~ rhs of a model}

\item{dat}{data.frame or tibble that contains data for the model formula}
}
\value{
Formula object that includes dummy terms and transforms for variables with missing data.
}
\description{
Missing values are dummy-encoded such that maximum-likelihood estimation
via lm or glm models will identify a value maximizing the model likelihood for missing observations.
}
\examples{

vars <- grep( "the_|rnd", colnames(toy_data), value=TRUE )
resp <- "resp"
df <- toy_data[,c(resp,vars[1:8])]
df$the_cubed[ sample.int(nrow(df), 5) ] <- NA
fo <- as.formula(df)
rfo <- rewrite_formula( fo, df )
rfo
glm( rfo, "binomial", df )
# saf(the_cubed) denotes the effect for the non-missing variables
# psu(the_cubed) denotes the maximum-likelihood imputation value for the missing data

}
