% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_construction.R
\name{simple_transforms}
\alias{simple_transforms}
\title{Function to evaluate and add simple variable tranformations}
\usage{
simple_transforms(
  dat,
  vars,
  transforms = c("sqrt", "cubert", "log", "zscore"),
  name_pattern = function(varname, transform_name) sprintf("\%s_\%s", varname,
    transform_name)
)
}
\arguments{
\item{dat}{A data.frame or tibble with data for each variable in vars}

\item{vars}{A character vector of variable names to process.}

\item{transforms}{A list of transformations to try. These can be of square root 'sqrt', cube root 'cubert', 
natural logarithm 'log', and z-Score to Normal distribution 'zscore'}

\item{name_pattern}{A function that takes two arguments 'varname' and 'transform_name' and generates a new
variable name under with the transformed data is stored. Default is 'varname + '_' + transform_name.}
}
\value{
list with two elements
\describe{
\item{data}{A dataset comprising additional columns with transformed variables, if they improve Normality.}
\item{transformations}{A list of lists with elements for variable, transformed variable, a formula term, Shapiro-Wilk W statistic, and the name of the transform.}
}
}
\description{
This function adds some simple transformations, as described in \link{following https://rcompanion.org/handbook/I_12.html}
to a dataset for a set of given variables. However, variables are only added if they improve Normality as measured by an
increase in Shapiro-Wilk W statistic.
}
