% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy data for feature selection}
\format{
A data frame with :
\describe{
\item{the_normal}{A standard normal variable}
\item{the_squared}{A standard normal variable that was squared}
\item{the_cubed}{A standard normal variable that was raised to the power of 3}
\item{the_exped}{A standard normal variable that was exp(x)}
\item{the_multi}{A variable that was sampled from a Gaussian mixture 
distribution}
\item{rnd..}{Random variables obtained form www.random.org}
\item{resp}{A binary response variable}
}
}
\source{
Generated by MCMC sampling as implemented in inst/raw-data/toy_data.R
}
\usage{
toy_data
}
\description{
This small dataset comprises toydata for running the feature
selection algorithms. It was generated using MCMC sampling from a JAGS
model and comprises the following special variables:
\describe{
\item{the_normal}{A standard normal variable}
\item{the_squared}{A standard normal variable that was squared}
\item{the_cubed}{A standard normal variable that was raised to the power of 3}
\item{the_exped}{A standard normal variable that was exp(x)}
\item{the_multi}{A variable that was sampled from a Gaussian mixture 
distribution}
\item{rnd..}{Random variables obtained form www.random.org}
\item{resp}{A binary response variable}
}
The model was a linear model for the logistic response using all described
variables, except the rnd... ones.
}
\keyword{datasets}
