% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_influential.R
\name{fn_infl_coefficients}
\alias{fn_infl_coefficients}
\alias{fn_predict_cox}
\alias{fn_predict_glm}
\alias{utils_influential}
\alias{influential_observations}
\title{Determine influential points for the current selection}
\usage{
fn_infl_coefficients(mod, ...)

fn_predict_cox(mod, dat, ...)

fn_predict_glm(mod, dat, ...)

influential_observations(
  dat,
  resp,
  selection,
  fn_train,
  fn_eval,
  fn_get_params,
  fn_predict,
  ...
)
}
\arguments{
\item{dat}{data.frame or tibble comprising data for model generation and validation.}

\item{resp}{Character string defining the response (lhs) of the model formula.}

\item{selection}{Character vector defining the list of variables for selection. Those are concatenated by '+' 
as the right hand side (rhs) of the modeling formula.}

\item{fn_train}{Function with signature function( dat, resp, selection, ... ) that returns a model or NULL in any other case on the given data dat.}

\item{fn_eval}{Function with signature function( dat, resp, selection, ... ) that returns a real number or NA in any other case, e.g. when model is NULL.}

\item{fn_get_params}{Function with signature function( model ) that returns an 1-row tibble of model parameters}

\item{fn_predict}{Function with signature function( model, data ) that returns an 1-column tibble of model predictions}
}
\value{
Returns a tibble with columns for observation (row), evaluation without observation i (ei), difference to full dataset, difference 
between prediction from full model vs model without observation i (y0, yi and deval), as well as differences between parameters (_dfbeta).
}
\description{
Determine influential points for the current selection
}
