% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reporting.R
\name{utils_reporting}
\alias{utils_reporting}
\alias{render_backward_summary}
\alias{render_forward_summary}
\alias{render_bidirectional_summary}
\alias{render_lrsearch_summary}
\alias{render_random_summary}
\alias{render_game_rank_summary}
\alias{render_variable_checks_summary}
\alias{render_model_calibration_normal}
\alias{render_model_calibration_binomial}
\alias{render_model_calibration_cox}
\title{Render a standard template reports for a given objects.}
\usage{
render_backward_summary(bwd, output_dir = NULL)

render_forward_summary(fwd, output_dir = NULL)

render_bidirectional_summary(bds, output_dir = NULL)

render_lrsearch_summary(lrs, output_dir = NULL)

render_random_summary(rnd, output_dir = NULL)

render_game_rank_summary(gmr, output_dir = NULL)

render_variable_checks_summary(vck, output_dir = NULL)

render_model_calibration_normal(ds, dat, resp, selection, k, output_dir = NULL)

render_model_calibration_binomial(
  ds,
  dat,
  resp,
  selection,
  k,
  output_dir = NULL
)

render_model_calibration_cox(ds, dat, resp, selection, k, u, output_dir = NULL)
}
\arguments{
\item{output_dir}{Output directory for the report to be saved to.}

\item{obj}{Algorithm results output from forward, backward and other selection methods}
}
\description{
Render a standard template reports for a given objects.
}
