% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{check_variable}
\alias{check_variable}
\title{Utility function that performs high-level checks for a variable in a dataset}
\usage{
check_variable(dat, var)
}
\arguments{
\item{dat}{data.frame or tibble comprising data for model generation and validation.}

\item{var}{Character indicating the variable to check.}
}
\value{

}
\description{
Utility function that performs high-level checks for a variable in a dataset
}
