% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{check_variables}
\alias{check_variables}
\title{Utility function that applies check_variable to a dataset}
\usage{
check_variables(dat, resp, vars)
}
\arguments{
\item{dat}{data.frame or tibble comprising data for model generation and validation.}

\item{resp}{Character string defining the response (lhs) of the model formula.}

\item{vars}{Character indicating the variables to check.}
}
\value{

}
\description{
Applies check_variable to each response and each variable in vars, and returns a merged dataset.
}
