% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy data for feature selection}
\format{
A data frame with :
\describe{
}
}
\source{
Generated by MCMC sampling
}
\usage{
toy_data
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
