% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy data for feature selection}
\format{
A data frame with :
\describe{
\item{USUBJID}{Toy sample ID}
\item{the_normal}{A standard normal variable}
\item{the_squared}{A standard normal variable that was squared}
\item{the_cubed}{A standard normal variable that was raised to the power of 3}
\item{the_power }{A standard normal variable that was raised to the power of 5}
\item{the_exped}{A standard normal variable that was exp(x)}
\item{the_multi}{A variable that was sampled from a Gaussian mixture distribution}
\item{resp}{A binary response variable}
\item{dys}{Simulated survival days}
\item{srv}{Binary survival flag}
\item{pi}{Survival probability}
\item{reg}{Numerical regression variable}
\item{rnd01}{Random variables obtained form www.random.org}
\item{rnd02}{Random variables obtained form www.random.org}
\item{rnd03}{Random variables obtained form www.random.org}
\item{rnd04}{Random variables obtained form www.random.org}
\item{rnd05}{Random variables obtained form www.random.org}
\item{rnd06}{Random variables obtained form www.random.org}
\item{rnd07}{Random variables obtained form www.random.org}
\item{rnd08}{Random variables obtained form www.random.org}
\item{rnd09}{Random variables obtained form www.random.org}
\item{rnd10}{Random variables obtained form www.random.org}
\item{rnd11}{Random variables obtained form www.random.org}
\item{rnd12}{Random variables obtained form www.random.org}
\item{rnd13}{Random variables obtained form www.random.org}
\item{rnd14}{Random variables obtained form www.random.org}
\item{rnd15}{Random variables obtained form www.random.org}
\item{rnd16}{Random variables obtained form www.random.org}
\item{rnd17}{Random variables obtained form www.random.org}
\item{rnd18}{Random variables obtained form www.random.org}
\item{rnd19}{Random variables obtained form www.random.org}
\item{rnd20}{Random variables obtained form www.random.org}
}
}
\source{
Generated by MCMC sampling as implemented in inst/raw-data/toy_data.R
}
\usage{
toy_data
}
\description{
This small dataset comprises toydata for running the feature
selection algorithms. It was generated using MCMC sampling from a JAGS
model and comprises the following special variables:
\describe{
\item{USUBJID}{Toy sample ID}
\item{the_normal}{A standard normal variable}
\item{the_squared}{A standard normal variable that was squared}
\item{the_cubed}{A standard normal variable that was raised to the power of 3}
\item{the_power }{A standard normal variable that was raised to the power of 5}
\item{the_exped}{A standard normal variable that was exp(x)}
\item{the_multi}{A variable that was sampled from a Gaussian mixture distribution}
\item{resp}{A binary response variable}
\item{dys}{Simulated survival days}
\item{srv}{Binary survival flag}
\item{pi}{Survival probability}
\item{reg}{Numerical regression variable}
\item{rnd01}{Random variables obtained form www.random.org}
\item{rnd02}{Random variables obtained form www.random.org}
\item{rnd03}{Random variables obtained form www.random.org}
\item{rnd04}{Random variables obtained form www.random.org}
\item{rnd05}{Random variables obtained form www.random.org}
\item{rnd06}{Random variables obtained form www.random.org}
\item{rnd07}{Random variables obtained form www.random.org}
\item{rnd08}{Random variables obtained form www.random.org}
\item{rnd09}{Random variables obtained form www.random.org}
\item{rnd10}{Random variables obtained form www.random.org}
\item{rnd11}{Random variables obtained form www.random.org}
\item{rnd12}{Random variables obtained form www.random.org}
\item{rnd13}{Random variables obtained form www.random.org}
\item{rnd14}{Random variables obtained form www.random.org}
\item{rnd15}{Random variables obtained form www.random.org}
\item{rnd16}{Random variables obtained form www.random.org}
\item{rnd17}{Random variables obtained form www.random.org}
\item{rnd18}{Random variables obtained form www.random.org}
\item{rnd19}{Random variables obtained form www.random.org}
\item{rnd20}{Random variables obtained form www.random.org}
}
The model was a linear model for the logistic response using all described
variables, except the rnd... ones.
}
\keyword{datasets}
