% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_influential.R
\name{fn_infl_coefficients}
\alias{fn_infl_coefficients}
\alias{fn_predict_cox}
\alias{fn_predict_glm}
\alias{utils_influential}
\alias{influential_observations}
\title{Determine influential points for the current selection}
\usage{
fn_infl_coefficients(mod, ...)

fn_predict_cox(mod, dat, ...)

fn_predict_glm(mod, dat, ...)

influential_observations(
  dat,
  resp,
  selection,
  fn_train,
  fn_eval,
  fn_get_params,
  fn_predict,
  c_out = 1.5,
  ...
)
}
\arguments{
\item{mod}{Model to obtain predictions from via predict(...)}

\item{...}{Further arguments passed to fn_train or fn_eval functions}

\item{dat}{data.frame or tibble comprising data for model generation and 
validation.}

\item{resp}{Character string defining the response (lhs) of the model formula.}

\item{selection}{Character vector defining the list of variables for 
selection. Those are concatenated by '+' 
as the right hand side (rhs) of the modeling formula.}

\item{fn_train}{Function with signature function( dat, resp, selection, ... ) 
that returns a model or NULL in any other case on the given data dat.}

\item{fn_eval}{Function with signature function( dat, resp, selection, ... ) 
that returns a real number or NA in any other case, e.g. when model is NULL.}

\item{fn_get_params}{Function with signature function( model ) that returns 
an 1-row tibble of model parameters}

\item{fn_predict}{Function with signature function( model, data ) that 
returns an 1-column tibble of model predictions}

\item{c_out}{Constant for the Robust Outlier Test determining the scaling of 
IQR to determine non-outlier range from Q1 - c_out x IQR to Q3 + c_out x IQR.}
}
\value{
Returns a tibble with columns for observation (row), evaluation 
without observation i (ei), difference to full dataset, difference 
between prediction from full model vs model without observation i (y0, yi and 
deval), as well as differences between parameters (_dfbeta).
}
\description{
Determine influential points for the current selection
}
\examples{
vars <- grep( "the_|rnd", colnames(toy_data), value=TRUE )
resp <- "resp"
gmr <- game_rank( dat = toy_data, resp = resp, vars = vars, 
                  fn_train = fn_train_binomial, fn_eval = fn_eval_binomial_auroc, 
                  m = 6L, dsi = c(1L,2L), maximize = TRUE, 
                  team_size = 3L, rounds = 10L, min_matches_per_var = 5L )
gmr$variable_ranking \%>\% as.data.frame
gmr_fsel <- gmr$game_rank_selection

ifo <- influential_observations( toy_data, resp, gmr_fsel, 
                                 fn_train_binomial, fn_eval_binomial_auroc, 
                                 fn_infl_coefficients, fn_predict_glm )
ifo 

}
