% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reporting.R
\name{utils_reporting}
\alias{utils_reporting}
\alias{render_std_template}
\alias{render_backward_summary}
\alias{render_forward_summary}
\alias{render_bidirectional_summary}
\alias{render_lrsearch_summary}
\alias{render_random_summary}
\alias{render_game_rank_summary}
\alias{render_variable_checks_summary}
\alias{render_model_calibration_normal}
\alias{render_model_calibration_binomial}
\alias{render_model_calibration_cox}
\title{Render a standard template reports for a given objects.}
\usage{
render_std_template(
  template_name,
  output_dir = NULL,
  params,
  output_file = NULL,
  envir = new.env(),
  ...
)

render_backward_summary(bwd, output_dir = NULL, output_file = NULL)

render_forward_summary(fwd, output_dir = NULL, output_file = NULL)

render_bidirectional_summary(bds, output_dir = NULL, output_file = NULL)

render_lrsearch_summary(lrs, output_dir = NULL, output_file = NULL)

render_random_summary(rnd, output_dir = NULL, output_file = NULL)

render_game_rank_summary(gmr, output_dir = NULL, output_file = NULL)

render_variable_checks_summary(vck, output_dir = NULL, output_file = NULL)

render_model_calibration_normal(
  ds,
  dat,
  resp,
  selection,
  k,
  output_dir = NULL,
  output_file = NULL
)

render_model_calibration_binomial(
  ds,
  dat,
  resp,
  selection,
  k,
  output_dir = NULL,
  output_file = NULL
)

render_model_calibration_cox(
  ds,
  dat,
  resp,
  selection,
  k,
  u,
  output_dir = NULL,
  output_file = NULL
)
}
\arguments{
\item{template_name}{File name of the standard template}

\item{output_dir}{Output directory to which to render}

\item{params}{The Rmd parameter list. Needs to match the expected parameter definitions of the Rmd template.}

\item{output_file}{Output filename (with or without suffix). Default NULL. If not NULL, it is appended to output_dir}

\item{...}{Further parameters passed to rmarkdown::render(...) function.}

\item{obj}{Algorithm results output from forward, backward and other selection methods}
}
\value{
render_... functions don't have return values. They generated the required output documents.
}
\description{
Render a standard template reports for a given objects.

Render a standard template report for a given object. (Internal function)
}
